feat: ModalContent 레이아웃 구현
