+++ components/src/lib/Modal/Modal.styled.tsx
@@ -2,31 +2,8 @@ import styled from "@emotion/styled";
 
 import { ModalProps } from "./Modal";
 
-const modalPositionVariants = {
-  bottom: {
-    left: "0",
-    bottom: "0",
-    width: "100%",
-    borderTopLeftRadius: "8px",
-    borderTopRightRadius: "8px",
-  },
-  center: {
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    width: "90%",
-    borderRadius: "8px",
-  },
-};
-
 export const StyledModal = styled.div<ModalProps>`
-  position: fixed;
-  padding: 24px 32px;
-  min-height: 216px;
-  background-color: #fff;
-  border: 1px solid red;
-  flex-direction: column;
-  justify-content: space-between;
-  ${({ position = "center" }) => modalPositionVariants[position]};
-  ${({ isOpen }) => (isOpen ? "display: flex" : "display: none")};
+  width: 100%;
+  height: 100%;
+  ${({ isOpen }) => (isOpen ? "display: block" : "display: none")};
 `;
