+++ components/src/lib/index.ts
@@ -1,5 +1,8 @@
 export { default as Modal } from "./Modal/Modal";
+export { default as ModalOverlay } from "./ModalOverlay/ModalOverlay";
+export { default as ModalContent } from "./ModalContent/ModalContent";
 export { default as ModalHeader } from "./ModalHeader/ModalHeader";
 export { default as ModalTitle } from "./ModalTitle/ModalTitle";
-export { default as ModalOverlay } from "./ModalOverlay/ModalOverlay";
+export { default as ModalCloseButton } from "./ModalCloseButton/ModalCloseButton";
 export { default as ModalBody } from "./ModalBody/ModalBody";
+export { default as ModalFooter } from "./ModalFooter/ModalFooter";
